split json ast emitter into separate bin
