+++ xdr-generator-rust/generator-definitions-json/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "generator-definitions-json"
+description = "Emit parsed XDR definitions as a JSON AST for downstream code generators"
+version = "0.1.0"
+edition = "2021"
+rust-version = "1.84.0"
+
+[[bin]]
+name = "generator-definitions-json"
+path = "src/main.rs"
+
+[dependencies]
+clap = { version = "4", features = ["derive"] }
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+xdr-parser = { path = "../xdr-parser" }
