+++ xdr-generator-rust/generator-definitions-json/src/tests/mod.rs
@@ -1,0 +1,2 @@
+mod cfg_filter;
+mod wire_size;
