+++ xdr-generator-rust/generator-definitions-json/src/tests/wire_size.rs
