+++ xdr-generator-rust/generator/Cargo.toml
@@ -13,8 +13,6 @@ path = "src/main.rs"
 askama = "0.14"
 clap = { version = "4", features = ["derive"] }
 heck = "0.5"
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
 xdr-parser = { path = "../xdr-parser" }
 
 [dev-dependencies]
