+++ xdr-generator-rust/generator/src/tests/mod.rs
@@ -1,5 +1,3 @@
-mod cfg_filter;
 mod generator;
 mod naming;
 mod types;
-mod wire_size;
